Rename colorpicker
